Copy the known_hosts file only if present

+++ docker-build/startup.sh
@@ -13,11 +13,13 @@ cp "/ssh_client_key" "/home/portmap/.ssh/ssh_client_key"
 chown portmap:portmap "/home/portmap/.ssh/ssh_client_key"
 chmod 600 "/home/portmap/.ssh/ssh_client_key"
 
-# Copy the known_hosts file in the home dir of the portmap user and set the
-# right permissions
-cp "/known_hosts" "/home/portmap/.ssh/known_hosts"
-chown portmap:portmap "/home/portmap/.ssh/known_hosts"
-chmod 644 "/home/portmap/.ssh/known_hosts"
+# Copy the known_hosts file (if exists) in the home dir of the portmap user and
+# set the right permissions
+if [ -f "/known_hosts" ]; then
+    cp "/known_hosts" "/home/portmap/.ssh/known_hosts"
+    chown portmap:portmap "/home/portmap/.ssh/known_hosts"
+    chmod 644 "/home/portmap/.ssh/known_hosts"
+fi
 
 # Run the client.sh script as the portmap user
 su portmap -c "/home/portmap/client.sh"
